chore: modularize shadow trader architecture
