+++ requirements.txt
@@ -1,3 +1,4 @@
 numpy>=1.24
 pandas>=1.5
 pygame>=2.5
+PyYAML>=6.0
